feat(mcp): separate orchestration from tool-crate and implement dual MCP server architecture
- Remove laravel-tool-crate monorepo directory (34 files)
- Migrate orchestration tools to app/Tools/Orchestration/ (13 tools)
- Create OrchestrationServer for Fragments Engine-specific tools
- Require laravel-tool-crate v0.2.2 from hollis-labs GitHub org
- Register both MCP servers (orchestration + tool-crate) in AppServiceProvider
- Update .mcp.json with both server configurations
- Fix tool names to match MCP pattern (dots → underscores)
- Update Laravel MCP to v0.2.1, Boost to v1.3.0

BREAKING CHANGES:
- Tool names changed from dot notation to underscore notation
  - json.query → json_query
  - orchestration.agents.list → orchestration_agents_list
  - etc.
- laravel-tool-crate now external package, not monorepo subdirectory

Architecture:
- tool-crate: Generic tools (json_query, text_search, file_read, text_replace, help)
- orchestration: Fragments Engine-specific (sprints, tasks, agents - 13 tools)
- Both exposed via MCP protocol for agent access
