+++ app/Console/Commands/OrchestrationMcp.php
@@ -1,0 +1,19 @@
+<?php
+
+namespace App\Console\Commands;
+
+use Illuminate\Console\Command;
+use Illuminate\Support\Facades\Artisan;
+use Symfony\Component\Console\Attribute\AsCommand;
+
+#[AsCommand('orchestration:mcp', 'Fragments Engine Orchestration MCP Server')]
+class OrchestrationMcp extends Command
+{
+    protected $signature = 'orchestration:mcp';
+    protected $description = 'Fragments Engine Orchestration MCP Server';
+
+    public function handle(): int
+    {
+        return Artisan::call('mcp:start', ['handle' => 'orchestration']);
+    }
+}
