+++ composer.json
@@ -37,6 +37,7 @@
     "require-dev": {
         "fakerphp/faker": "^1.23",
         "hollis-labs/laravel-console": "^0.1.1",
+        "hollis-labs/laravel-tool-crate": "^0.2.1",
         "laravel/boost": "^1.0",
         "laravel/pail": "^1.2.2",
         "laravel/pint": "^1.13",
@@ -51,8 +52,7 @@
         "psr-4": {
             "App\\": "app/",
             "Database\\Factories\\": "database/factories/",
-            "Database\\Seeders\\": "database/seeders/",
-            "HollisLabs\\ToolCrate\\": "laravel-tool-crate/src/"
+            "Database\\Seeders\\": "database/seeders/"
         }
     },
     "autoload-dev": {
@@ -130,9 +130,13 @@
             "type": "composer",
             "url": "https://composer.fluxui.dev"
         },
-        "0": {
+        "laravel-console": {
             "type": "vcs",
             "url": "https://github.com/hollis-labs/laravel-console"
+        },
+        "laravel-tool-crate": {
+            "type": "vcs",
+            "url": "https://github.com/hollis-labs/laravel-tool-crate"
         }
     }
 }
