--- laravel-tool-crate/composer.json
@@ -1,33 +1,0 @@
-{
-  "name": "hollis-labs/laravel-tool-crate",
-  "description": "Opinionated local dev MCP server for Laravel (official laravel/mcp), with a context-lean help layer and CLI-parallel tools.",
-  "type": "library",
-  "license": "MIT",
-  "autoload": {
-    "psr-4": {
-      "HollisLabs\\ToolCrate\\": "src/"
-    }
-  },
-  "autoload-dev": {
-    "psr-4": {
-      "HollisLabs\\ToolCrate\\Tests\\": "tests/"
-    }
-  },
-  "require": {
-    "php": ">=8.2",
-    "illuminate/support": "^10.0|^11.0|^12.0",
-    "laravel/mcp": "^1.0|^0.1|dev-main",
-    "symfony/process": "^7.0",
-    "symfony/finder": "^7.0",
-    "spatie/laravel-package-tools": "^1.16"
-  },
-  "extra": {
-    "laravel": {
-      "providers": [
-        "HollisLabs\\ToolCrate\\ToolCrateServiceProvider"
-      ]
-    }
-  },
-  "minimum-stability": "stable",
-  "prefer-stable": true
-}
