--- laravel-tool-crate/config/tool-crate.php
@@ -1,64 +1,0 @@
-<?php
-
-return [
-    'enabled_tools' => [
-        'json.query'                 => true,
-        'text.search'                => true,
-        'file.read'                  => true,
-        'text.replace'               => true,
-        'help.index'                 => true,
-        'help.tool'                  => true,
-        'orchestration.agents.list'  => true,
-        'orchestration.agents.detail' => true,
-        'orchestration.agents.save' => true,
-        'orchestration.agents.status' => true,
-        'orchestration.tasks.list'   => true,
-        'orchestration.sprints.list' => true,
-        'orchestration.tasks.detail' => true,
-        'orchestration.tasks.assign' => true,
-        'orchestration.tasks.status' => true,
-        'orchestration.sprints.detail' => true,
-        'orchestration.sprints.save' => true,
-        'orchestration.sprints.status' => true,
-        'orchestration.sprints.attach_tasks' => true,
-    ],
-    'priority_tools' => [
-        'json.query',
-        'text.search',
-        'file.read',
-        'orchestration.agents.list',
-        'orchestration.agents.save',
-        'orchestration.tasks.list',
-        'orchestration.tasks.detail',
-        'orchestration.sprints.save',
-    ],
-    'categories' => [
-        'JSON & Data' => ['json.query', 'table.query'],
-        'Text Ops'    => ['text.search', 'text.replace'],
-        'Files'       => ['file.read'],
-        'Help'        => ['help.index', 'help.tool'],
-        'Orchestration' => [
-            'orchestration.agents.list',
-            'orchestration.agents.detail',
-            'orchestration.agents.save',
-            'orchestration.agents.status',
-            'orchestration.tasks.list',
-            'orchestration.sprints.list',
-            'orchestration.tasks.detail',
-            'orchestration.tasks.assign',
-            'orchestration.tasks.status',
-            'orchestration.sprints.detail',
-            'orchestration.sprints.save',
-            'orchestration.sprints.status',
-            'orchestration.sprints.attach_tasks',
-        ],
-    ],
-    'orchestration' => [
-        'agent_model' => 'App\\Models\\AgentProfile',
-        'sprint_model' => 'App\\Models\\Sprint',
-        'work_item_model' => 'App\\Models\\WorkItem',
-        'task_service' => 'App\\Services\\TaskOrchestrationService',
-        'sprint_service' => 'App\\Services\\SprintOrchestrationService',
-        'agent_service' => 'App\\Services\\AgentOrchestrationService',
-    ],
-];
