--- laravel-tool-crate/phpunit.xml.dist
@@ -1,8 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<phpunit bootstrap="vendor/autoload.php">
-  <testsuites>
-    <testsuite name="Package Test Suite">
-      <directory suffix="Test.php">tests</directory>
-    </testsuite>
-  </testsuites>
-</phpunit>
