--- laravel-tool-crate/routes/ai.php
@@ -1,6 +1,0 @@
-<?php
-
-use Laravel\Mcp\Facades\Mcp;
-use HollisLabs\ToolCrate\Servers\ToolCrateServer;
-
-Mcp::local('tool-crate', ToolCrateServer::class);
