--- laravel-tool-crate/src/Tools/Contracts/SummarizesTool.php
@@ -1,12 +1,0 @@
-<?php
-
-namespace HollisLabs\ToolCrate\Tools\Contracts;
-
-interface SummarizesTool
-{
-    public static function summaryName(): string;
-    public static function summaryTitle(): string;
-    public static function summaryDescription(): string;
-    /** @return array<string,string> */
-    public static function schemaSummary(): array;
-}
