--- laravel-tool-crate/src/Tools/Orchestration/TaskStatusTool.php
@@ -1,81 +1,0 @@
-<?php
-
-namespace HollisLabs\ToolCrate\Tools\Orchestration;
-
-use HollisLabs\ToolCrate\Support\Orchestration\ModelResolver;
-use HollisLabs\ToolCrate\Tools\Contracts\SummarizesTool;
-use Illuminate\JsonSchema\JsonSchema;
-use Illuminate\Support\Arr;
-use Laravel\Mcp\Request;
-use Laravel\Mcp\Response;
-use Laravel\Mcp\Server\Tool;
-
-class TaskStatusTool extends Tool implements SummarizesTool
-{
-    protected string $name = 'orchestration.tasks.status';
-    protected string $title = 'Update orchestration task status';
-    protected string $description = 'Change delegation status for a work item and sync the active assignment.';
-
-    public function schema(JsonSchema $schema): array
-    {
-        return [
-            'task' => $schema->string()->required()->description('Task UUID or delegation task code'),
-            'status' => $schema->string()->enum([
-                'unassigned', 'assigned', 'in_progress', 'blocked', 'completed', 'cancelled',
-            ])->required(),
-            'assignment_status' => $schema->string()->enum([
-                'unassigned', 'assigned', 'in_progress', 'blocked', 'completed', 'cancelled',
-            ])->description('Override assignment status; defaults to status value'),
-            'note' => $schema->string()->description('Optional note stored alongside history'),
-            'assignments_limit' => $schema->integer()->minimum(1)->maximum(20)->default(10),
-            'include_history' => $schema->boolean()->default(true),
-        ];
-    }
-
-    public function handle(Request $request): Response
-    {
-        $validated = $request->validate([
-            'task' => ['required', 'string'],
-            'status' => ['required', 'string'],
-            'assignment_status' => ['nullable', 'string'],
-            'note' => ['nullable', 'string'],
-            'assignments_limit' => ['nullable', 'integer'],
-            'include_history' => ['nullable', 'boolean'],
-        ]);
-
-        $service = ModelResolver::resolveService('task_service', 'App\\Services\\TaskOrchestrationService');
-
-        $task = $service->updateStatus($validated['task'], $validated['status'], Arr::only($validated, ['assignment_status', 'note']));
-
-        $detail = $service->detail($task, [
-            'assignments_limit' => $validated['assignments_limit'] ?? 10,
-            'include_history' => $validated['include_history'] ?? true,
-        ]);
-
-        return Response::json($detail);
-    }
-
-    public static function summaryName(): string
-    {
-        return 'orchestration.tasks.status';
-    }
-
-    public static function summaryTitle(): string
-    {
-        return 'Update orchestration task status';
-    }
-
-    public static function summaryDescription(): string
-    {
-        return 'Set delegation status and reflect it on the most recent assignment.';
-    }
-
-    public static function schemaSummary(): array
-    {
-        return [
-            'task' => 'Task UUID or code',
-            'status' => 'unassigned|assigned|in_progress|blocked|completed|cancelled',
-            'note' => 'optional note captured in history',
-        ];
-    }
-}
